refactor: txProposalCreate and txProposalSend expecting txHex
