+++ src/api/txProposalCreate.ts
@@ -6,74 +6,40 @@
  */
 
 import { ServerlessMysql } from 'serverless-mysql';
+import { APIGatewayProxyHandler } from 'aws-lambda';
 import 'source-map-support/register';
 import { v4 as uuidv4 } from 'uuid';
 import Joi from 'joi';
 
 import { ApiError } from '@src/api/errors';
-import { getWalletBalances, maybeRefreshWalletConstants, walletIdProxyHandler } from '@src/commons';
+import { maybeRefreshWalletConstants, walletIdProxyHandler } from '@src/commons';
 import {
   addTxProposalOutputs,
   createTxProposal,
-  getUnusedAddresses,
+  // getUnusedAddresses,
   getUtxos,
   getWallet,
   getWalletAddressDetail,
-  getWalletSortedValueUtxos,
   markUtxosWithProposalId,
+  addTxProposalTokenInfo,
 } from '@src/db';
 import {
   AddressInfo,
-  Balance,
   IWalletInput,
   IWalletOutput,
-  TokenBalanceMap,
   DbTxOutput,
-  WalletTokenBalance,
-  InputSelectionAlgo,
-  IWalletInsufficientFunds,
 } from '@src/types';
 import { closeDbAndGetError } from '@src/api/utils';
-import { arrayShuffle, closeDbConnection, getDbConnection, getUnixTimestamp } from '@src/utils';
+import { closeDbConnection, getDbConnection, getUnixTimestamp } from '@src/utils';
 import hathorLib from '@hathor/wallet-lib';
 
 const mysql = getDbConnection();
 
 const bodySchema = Joi.object({
-  outputs: Joi.array()
-    .items(
-      Joi.object({
-        address: Joi.string()
-          .alphanum()
-          .required(),
-        value: Joi.number()
-          .integer()
-          .positive()
-          .required(),
-        token: Joi.string()
-          .alphanum(),
-        timelock: Joi.number()
-          .integer()
-          .positive()
-          .optional()
-          .allow(null),
-      }),
-    )
-    .min(1)
+  id: Joi.string()
     .required(),
-  inputs: Joi.array()
-    .items(
-      Joi.object({
-        txId: Joi.string()
-          .alphanum()
-          .required(),
-        index: Joi.number()
-          .integer()
-          .required()
-          .min(0),
-      }),
-    ),
-  inputSelectionAlgo: Joi.string(),
+  txHex: Joi.string().alphanum(),
+  send: Joi.boolean().default(false),
 });
 
 /*
@@ -107,27 +73,40 @@ export const create = walletIdProxyHandler(async (walletId, event) => {
   }
 
   const body = value;
+  const walletId = body.id;
+  const tx = hathorLib.helpersUtils.createTxFromHex(body.txHex, new hathorLib.Network(process.env.NETWORK));
 
-  if (body.outputs.length > hathorLib.transaction.getMaxOutputsConstant()) {
-    return closeDbAndGetError(mysql, ApiError.TOO_MANY_OUTPUTS, { outputs: body.outputs.length });
+  if (tx.outputs.length > hathorLib.transaction.getMaxOutputsConstant()) {
+    return closeDbAndGetError(mysql, ApiError.TOO_MANY_OUTPUTS, { outputs: tx.outputs.length });
   }
 
-  const outputs: IWalletOutput[] = body.outputs;
-  const inputs: IWalletInput[] = body.inputs;
-
-  const inputSelectionAlgo = (function getInputAlgoFromBody() {
-    if (!body.inputSelectionAlgo) {
-      return InputSelectionAlgo.USE_LARGER_UTXOS;
+  const getToken = (tokenData, tokens) => {
+    if (tokenData === 0) {
+      return '00';
     }
 
-    return InputSelectionAlgo[body.inputSelectionAlgo];
-  }());
+    if (!tokenData) {
+      return null;
+    }
 
-  if (!inputSelectionAlgo) {
-    return closeDbAndGetError(mysql, ApiError.INVALID_SELECTION_ALGORITHM);
-  }
+    return tokens[hathorLib.wallet.getTokenIndex(tokenData) - 1];
+  };
+
+  const tokens = tx.tokens;
+  const outputs: IWalletOutput[] = tx.outputs.map((output) => ({
+    address: output.address.base58,
+    value: output.value,
+    token: getToken(output.tokenData, tokens),
+    tokenData: output.tokenData,
+    timelock: output.timelock,
+  }));
+  const inputs: IWalletInput[] = tx.inputs.map((input) => ({
+    txId: input.hash,
+    index: input.index,
+  }));
 
   const status = await getWallet(mysql, walletId);
+
   if (!status) {
     return closeDbAndGetError(mysql, ApiError.WALLET_NOT_FOUND);
   }
@@ -137,78 +116,35 @@ export const create = walletIdProxyHandler(async (walletId, event) => {
   }
 
   const now = getUnixTimestamp();
-  const outputsBalance = getOutputsBalance(outputs, now);
-
-  // check if wallet's balances are enough for the request
-  const balances = await getWalletBalances(mysql, now, walletId, outputsBalance.getTokens());
-  const insufficientFunds = checkWalletFunds(balances, outputsBalance);
-  if (insufficientFunds.length > 0) {
-    return closeDbAndGetError(mysql, ApiError.INSUFFICIENT_FUNDS, { insufficient: insufficientFunds });
-  }
 
   // fetch the utxos that will be used
-  let inputUtxos = [];
-  if (inputs && inputs.length > 0) {
-    inputUtxos = await getUtxos(mysql, inputs);
+  const inputUtxos = await getUtxos(mysql, inputs);
+  const missing = checkMissingUtxos(inputs, inputUtxos);
 
-    const missing = checkMissingUtxos(inputs, inputUtxos);
+  if (missing.length > 0) {
+    return closeDbAndGetError(mysql, ApiError.INPUTS_NOT_FOUND, { missing });
+  }
 
-    if (missing.length > 0) {
-      return closeDbAndGetError(mysql, ApiError.INPUTS_NOT_FOUND, { missing });
-    }
-
-    // check if inputs sent by user are not part of another tx proposal
-    if (checkUsedUtxos(inputUtxos)) {
-      return closeDbAndGetError(mysql, ApiError.INPUTS_ALREADY_USED);
-    }
-  } else {
-    for (const [tokenId, tokenBalance] of outputsBalance.iterator()) {
-      const utxos = await getUtxosForTokenBalance(mysql, inputSelectionAlgo, walletId, tokenId, tokenBalance);
-      inputUtxos.push(...utxos);
-    }
+  // check if inputs sent by user are not part of another tx proposal
+  if (checkUsedUtxos(inputUtxos)) {
+    return closeDbAndGetError(mysql, ApiError.INPUTS_ALREADY_USED);
   }
 
   if (inputUtxos.length > hathorLib.transaction.getMaxInputsConstant()) {
     return closeDbAndGetError(mysql, ApiError.TOO_MANY_INPUTS, { inputs: inputUtxos.length });
   }
 
-  // the difference between inputs and outputs will be the change
-  const inputsBalance = getInputsBalance(inputUtxos);
-  const diff = TokenBalanceMap.merge(outputsBalance, inputsBalance);
-
-  // Make sure diff is 0 or lower, which means inputs sum is greater than (or equal to) outputs sum.
-  // This should only happen when we receive the inputs from user and he didn't select enough inputs.
-  const insufficientInputs = [];
-  for (const [token, tokenBalance] of diff.iterator()) {
-    if (tokenBalance.total() > 0) insufficientInputs.push(token);
-  }
-  if (insufficientInputs.length > 0) {
-    return closeDbAndGetError(mysql, ApiError.INSUFFICIENT_INPUTS, { insufficient: insufficientInputs });
-  }
-
-  const addresses = await getUnusedAddresses(mysql, walletId);
-  const changeOutputs = getChangeOutputs(diff, addresses);
-
-  const finalOutputs = outputs.concat(changeOutputs);
-
-  // we also need to do this check here, as we may have added change outputs
-  if (finalOutputs.length > hathorLib.transaction.getMaxOutputsConstant()) {
-    return closeDbAndGetError(mysql, ApiError.TOO_MANY_OUTPUTS, { outputs: finalOutputs.length });
-  }
-
-  /**
-   * We shuffle the array to prevent the change address from always being the last output so we can give some more
-   * privacy to the user
-   */
-  arrayShuffle(finalOutputs);
-
   // mark utxos with tx-proposal id
   // XXX should this be done atomically?
   const txProposalId = uuidv4();
   markUtxosWithProposalId(mysql, txProposalId, inputUtxos);
 
-  await createTxProposal(mysql, txProposalId, walletId, now);
-  await addTxProposalOutputs(mysql, txProposalId, finalOutputs);
+  if (tx.version === hathorLib.constants.CREATE_TOKEN_TX_VERSION) {
+    await addTxProposalTokenInfo(mysql, txProposalId, tx.name, tx.symbol);
+  }
+
+  await createTxProposal(mysql, txProposalId, walletId, tx.version, now);
+  await addTxProposalOutputs(mysql, txProposalId, outputs);
 
   await closeDbConnection(mysql);
 
@@ -223,179 +159,17 @@ export const create = walletIdProxyHandler(async (walletId, event) => {
 
   const retInputs = await Promise.all(inputPromises);
 
-  // We need to return the tokens array as it's needed to assemble the tx for mining
-  const tokens = Array.from(new Set([...outputsBalance.getTokens()]));
-
   return {
     statusCode: 201,
     body: JSON.stringify({
       success: true,
       txProposalId,
       inputs: retInputs,
-      outputs: finalOutputs,
+      outputs,
       tokens,
     }),
   };
 });
-
-/**
- * Calculates the total balance for the outputs.
- *
- * @param outputs - List of outputs
- * @param now - Current timestamp
- * @returns A balance map merging all outputs
- */
-export const getOutputsBalance = (outputs: IWalletOutput[], now: number): TokenBalanceMap => {
-  let outputsBalance = null;
-  for (const output of outputs) {
-    const decoded = { type: 'P2PKH', address: output.address, timelock: output.timelock };
-    // take advantage of TokenBalanceMap.fromTxOuput
-    const txOutput = {
-      decoded,
-      value: output.value,
-      token: output.token,
-      // This is being done in https://github.com/HathorNetwork/hathor-wallet-service/pull/13
-      token_data: 0,
-      script: null,
-      spent_by: null,
-      locked: output.timelock > now,
-    };
-    outputsBalance = TokenBalanceMap.merge(outputsBalance, TokenBalanceMap.fromTxOutput(txOutput));
-  }
-  return outputsBalance;
-};
-
-/**
- * Calculates the total balance for the input UTXOs.
- *
- * @param inputUtxos - List of input UTXOs
- * @returns A balance map merging all input UTXOs
- */
-export const getInputsBalance = (inputUtxos: DbTxOutput[]): TokenBalanceMap => {
-  let inputsBalance = null;
-  for (const utxo of inputUtxos) {
-    const decoded = { type: 'P2PKH', address: utxo.address, timelock: utxo.timelock };
-    // take advantage of TokenBalanceMap.fromTxInput
-    const txInput = {
-      decoded,
-      tx_id: utxo.txId,
-      index: utxo.index,
-      value: utxo.value,
-      token: utxo.tokenId,
-      // TODO not handling authorities
-      token_data: 0,
-      script: null,
-    };
-    inputsBalance = TokenBalanceMap.merge(inputsBalance, TokenBalanceMap.fromTxInput(txInput));
-  }
-  return inputsBalance;
-};
-
-/**
- * Create the change outputs, given the difference between inputs and outputs.
- *
- * @remarks
- * Balances should either be zero or negative. A positive balance would indicate that the sum of
- * outputs is greater than the inputs, which would make an invalid transaction.
- *
- * @param diff - The difference between outputs and inputs
- * @returns The change outputs
- */
-export const getChangeOutputs = (diff: TokenBalanceMap, addresses: string[]): IWalletOutput[] => {
-  const changeOutputs = [];
-  let addressToUse = 0;
-  for (const [token, balance] of diff.iterator()) {
-    if (balance.total() < 0) {
-      changeOutputs.push({
-        address: addresses[addressToUse++],
-        value: Math.abs(balance.total()),
-        token,
-        timelock: null,
-      });
-
-      if (addressToUse >= addresses.length) {
-        // this treats an unlikely case, where we have more change outputs than addresses. In this case,
-        // we will repeat some addresses. Ideally, we should just generate more, but it's so unlikely
-        // that this happens that we can handle it later
-        addressToUse = 0;
-      }
-    }
-  }
-  return changeOutputs;
-};
-
-/**
- * Select the UTXOs to be spent, given the token balance.
- *
- * @param _mysql - The database connection
- * @param inputSelectionAlgo - The input selection algorithm
- * @param walletId - The wallet id
- * @param tokenId - The token id
- * @param tokenBalance - Balance for the queried token
- * @returns A list of UTXOs that sum at least the requested balance
- */
-const getUtxosForTokenBalance = async (
-  _mysql: ServerlessMysql,
-  inputSelectionAlgo: InputSelectionAlgo,
-  walletId: string,
-  tokenId: string,
-  tokenBalance: Balance,
-): Promise<DbTxOutput[]> => {
-  switch (inputSelectionAlgo) {
-    case InputSelectionAlgo.USE_LARGER_UTXOS:
-    default:
-      return useLargerUtxos(_mysql, walletId, tokenId, tokenBalance.total());
-  }
-};
-
-export const useLargerUtxos = async (
-  _mysql: ServerlessMysql,
-  walletId: string,
-  tokenId: string,
-  balance: number,
-): Promise<DbTxOutput[]> => {
-  const finalUtxos: DbTxOutput[] = [];
-
-  let remainingBalance = balance;
-  const valueUtxos = await getWalletSortedValueUtxos(_mysql, walletId, tokenId);
-  for (const utxo of valueUtxos) {
-    remainingBalance -= utxo.value;
-    finalUtxos.push(utxo);
-    if (remainingBalance <= 0) break;
-  }
-
-  return finalUtxos;
-};
-
-/**
- * Check if the wallet has the required amount for each token.
- *
- * @remarks
- * The check is only done using the wallet's unlocked tokens.
- *
- * @param walletBalances - The wallet's balance for all requested tokens
- * @param outputsBalance - The amount requested for each token
- * @returns A list of tokens whose requested value is larger than the wallet's available balance
- */
-export const checkWalletFunds = (walletBalances: WalletTokenBalance[], outputsBalance: TokenBalanceMap): IWalletInsufficientFunds[] => {
-  const insufficientFunds = [];
-  const missingTokens = new Set(outputsBalance.getTokens());
-  for (const balance of walletBalances) {
-    const token = balance.token;
-    missingTokens.delete(token.id);
-    const requested = outputsBalance.get(token.id).total();
-    const available = balance.balance.unlockedAmount;
-    if (requested > available) {
-      // unlocked tokens are not enough
-      insufficientFunds.push({ tokenId: token.id, requested, available });
-    }
-  }
-  for (const tokenId of missingTokens) {
-    // these tokens don't have any balance in the wallet
-    insufficientFunds.push({ tokenId, requested: outputsBalance.get(tokenId).total(), available: 0 });
-  }
-  return insufficientFunds;
-};
 
 /**
  * Confirm that all inputs requested by the user have been fetched.
