--- src/api/txProposalCreate2.ts
@@ -1,196 +1,0 @@
-import { APIGatewayProxyHandler } from 'aws-lambda';
-import 'source-map-support/register';
-import { v4 as uuidv4 } from 'uuid';
-import Joi from 'joi';
-
-import { ApiError } from '@src/api/errors';
-import { maybeRefreshWalletConstants } from '@src/commons';
-import {
-  addTxProposalOutputs,
-  createTxProposal,
-  // getUnusedAddresses,
-  getUtxos,
-  getWallet,
-  getWalletAddressDetail,
-  markUtxosWithProposalId,
-  addTxProposalTokenInfo,
-} from '@src/db';
-import {
-  AddressInfo,
-  IWalletInput,
-  IWalletOutput,
-  DbTxOutput,
-} from '@src/types';
-import { closeDbAndGetError } from '@src/api/utils';
-import { closeDbConnection, getDbConnection, getUnixTimestamp } from '@src/utils';
-import hathorLib from '@hathor/wallet-lib';
-
-const mysql = getDbConnection();
-
-const bodySchema = Joi.object({
-  id: Joi.string()
-    .required(),
-  txHex: Joi.string().alphanum(),
-  send: Joi.boolean().default(false),
-});
-
-/*
- * Create a tx-proposal.
- *
- * This lambda is called by API Gateway on POST /txproposals
- */
-export const create: APIGatewayProxyHandler = async (event) => {
-  await maybeRefreshWalletConstants(mysql);
-
-  const eventBody = (function parseBody(body) {
-    try {
-      return JSON.parse(body);
-    } catch (e) {
-      return null;
-    }
-  }(event.body));
-
-  const { value, error } = bodySchema.validate(eventBody, {
-    abortEarly: false, // We want it to return all the errors not only the first
-    convert: false, // We want it to be strict with the parameters and not parse a string as integer
-  });
-
-  if (error) {
-    const details = error.details.map((err) => ({
-      message: err.message,
-      path: err.path,
-    }));
-
-    return closeDbAndGetError(mysql, ApiError.INVALID_PAYLOAD, { details });
-  }
-
-  const body = value;
-  const walletId = body.id;
-  const tx = hathorLib.helpersUtils.createTxFromHex(body.txHex, new hathorLib.Network(process.env.NETWORK));
-
-  if (tx.outputs.length > hathorLib.transaction.getMaxOutputsConstant()) {
-    return closeDbAndGetError(mysql, ApiError.TOO_MANY_OUTPUTS, { outputs: tx.outputs.length });
-  }
-
-  const tokens = tx.tokens;
-  const outputs: IWalletOutput[] = tx.outputs.map((output) => ({
-    address: output.address.base58,
-    value: output.value,
-    token: tokens[hathorLib.wallet.getTokenIndex(output.tokenData) - 1],
-    tokenData: output.tokenData,
-    timelock: output.timelock,
-  }));
-  const inputs: IWalletInput[] = tx.inputs.map((input) => ({
-    txId: input.hash,
-    index: input.index,
-  }));
-
-  console.log('Outputs: ', outputs);
-  console.log('Inputs: ', inputs);
-
-  console.log('getWallet', getWallet);
-  /*
-  const status = await getWallet(mysql, walletId);
-
-  if (!status) {
-    return closeDbAndGetError(mysql, ApiError.WALLET_NOT_FOUND);
-  }
-
-  if (!status.readyAt) {
-    return closeDbAndGetError(mysql, ApiError.WALLET_NOT_READY);
-  }
-  */
-
-  const now = getUnixTimestamp();
-
-  // fetch the utxos that will be used
-  const inputUtxos = await getUtxos(mysql, inputs);
-  const missing = checkMissingUtxos(inputs, inputUtxos);
-
-  if (missing.length > 0) {
-    return closeDbAndGetError(mysql, ApiError.INPUTS_NOT_FOUND, { missing });
-  }
-
-  // check if inputs sent by user are not part of another tx proposal
-  if (checkUsedUtxos(inputUtxos)) {
-    return closeDbAndGetError(mysql, ApiError.INPUTS_ALREADY_USED);
-  }
-
-  if (inputUtxos.length > hathorLib.transaction.getMaxInputsConstant()) {
-    return closeDbAndGetError(mysql, ApiError.TOO_MANY_INPUTS, { inputs: inputUtxos.length });
-  }
-
-  // mark utxos with tx-proposal id
-  // XXX should this be done atomically?
-  const txProposalId = uuidv4();
-  markUtxosWithProposalId(mysql, txProposalId, inputUtxos);
-
-  if (tx.version === hathorLib.constants.CREATE_TOKEN_TX_VERSION) {
-    await addTxProposalTokenInfo(mysql, txProposalId, tx.name, tx.symbol);
-  }
-
-  await createTxProposal(mysql, txProposalId, walletId, now);
-  await addTxProposalOutputs(mysql, txProposalId, outputs);
-
-  await closeDbConnection(mysql);
-
-  const inputPromises = inputUtxos.map(async (utxo) => {
-    const addressDetail: AddressInfo = await getWalletAddressDetail(mysql, walletId, utxo.address);
-    // XXX We should store in address table the path of the address, not the index
-    // For now we return the hardcoded path with only the address index as variable
-    // The client will be prepared to receive any path when we add this in the service in the future
-    const addressPath = `m/44'/${hathorLib.constants.HATHOR_BIP44_CODE}'/0'/0/${addressDetail.index}`;
-    return { txId: utxo.txId, index: utxo.index, addressPath };
-  });
-
-  const retInputs = await Promise.all(inputPromises);
-
-  return {
-    statusCode: 201,
-    body: JSON.stringify({
-      success: true,
-      txProposalId,
-      inputs: retInputs,
-      outputs,
-      tokens,
-    }),
-  };
-};
-
-/**
- * Confirm that all inputs requested by the user have been fetched.
- *
- * @param inputs - List of inputs sent by the user
- * @param utxos - List of UTXOs retrieved from database
- * @returns A list with the missing UTXOs, if any
- */
-export const checkMissingUtxos = (inputs: IWalletInput[], utxos: DbTxOutput[]): IWalletInput[] => {
-  if (inputs.length === utxos.length) return [];
-
-  const remaining = new Set(inputs.map((input) => [input.txId, input.index]));
-  for (const utxo of utxos) {
-    remaining.delete([utxo.txId, utxo.index]);
-  }
-
-  const missing = [];
-  for (const utxo of remaining) {
-    missing.push({ txId: utxo[0], index: utxo[1] });
-  }
-  return missing;
-};
-
-/**
- * Confirm that the inputs requested by the user are not already being used on another TxProposal
- *
- * @param utxos - List of UTXOs retrieved from database
- * @returns A list with the missing UTXOs, if any
- */
-export const checkUsedUtxos = (utxos: DbTxOutput[]): boolean => {
-  for (let x = 0; x < utxos.length; x++) {
-    if (utxos[x].txProposalId) {
-      return true;
-    }
-  }
-
-  return false;
-};
